+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Adding label value validation for `Cluster` CR for non-version labels.
+- Adding label key validation for `Cluster` CR for `giantswarm.io` labels.
 
 ## [2.6.0] - 2020-12-07